Allow minimal resource loading and avoid partial refreshes.
Introduce a configuration flag that allows a user to optionally
cause a full resource load (off by default). This lets an admin
to configure the plugin to only load the minimum resource data.

Fixes #15

Change the needs refresh check to be based on the whole refresh
cycle and not the caching state.

Fixes #16

Add a node attribute, plugin:last-refresh, that tracks the time
when the node was loaded from the cache.

Fixes #14